+++ src/ninja/mbedded/ninjaterm/model/terminal/txRx/display/Display.java
@@ -2,6 +2,7 @@ package ninja.mbedded.ninjaterm.model.terminal.txRx.display;
 
 import javafx.beans.property.SimpleBooleanProperty;
 import javafx.beans.property.SimpleDoubleProperty;
+import javafx.beans.property.SimpleIntegerProperty;
 import javafx.beans.property.SimpleObjectProperty;
 
 /**
@@ -31,9 +32,12 @@ public class Display {
         SEND_TX_CHARS_ON_ENTER
     }
 
+    public final int DEFAULT_BUFFER_SIZE_CHARS = 10000;
+
     public SimpleBooleanProperty localTxEcho = new SimpleBooleanProperty(false);
     public SimpleObjectProperty<LayoutOptions> selectedLayoutOption = new SimpleObjectProperty<>();
     public SimpleObjectProperty<TxCharSendingOptions> selTxCharSendingOption = new SimpleObjectProperty<>();
+    public SimpleIntegerProperty bufferSizeChars = new SimpleIntegerProperty(DEFAULT_BUFFER_SIZE_CHARS);
 
     public SimpleBooleanProperty wrappingEnabled = new SimpleBooleanProperty(false);
     public SimpleDoubleProperty wrappingWidth = new SimpleDoubleProperty(800.0);